pyproject.toml
[project]
name = "tcai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.29",
    "djangorestframework>=3.16.1",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
